Add style config files, enable on build
Initially add it to the composite build because most people will be using either Python or Rust for active development and not running this build script repeatedly. You will in cases where you're testing the bindings, and if it causes a hinderance in those situations we'll remove it. But for now, it's a nice middle ground.