Re-add qubit to cluster if multi-op unentangled it
When applying multi qubit operations sometimes each individual one would make the qubit drop out of entanglement, which would then trigger removal during iteration. This is a quick fix to that, but it needs some careful considerations about when qubits remove themselves from things.

It may be easier to simply revert to the high level to do it for clusters.